Prepare for launch (#44)
* Use OkHttp

* Remove GlobalScope usage

* [MegaLinter] Apply linters fixes

* Migrate to postgresql

* Raise admin commands priority

* Dockerize

* [MegaLinter] Apply linters fixes

* Fix trendy-friendy

* Port preaccelerator Xlsx parser

* Fix upload phone numbers ending

* Hide steps other than first one

* Fix InstantSerializer

* Change token location

* Disable upload trends button

* Fix UnableToAddAdmin message

* Put edits in runCatching

+++ server/src/main/kotlin/ru/spbstu/application/auth/entities/User.kt
@@ -7,8 +7,8 @@ data class User(
     override val phoneNumber: PhoneNumber,
     override val avatar: Avatar,
     override val occupation: Occupation,
-    override val availableStepsCount: Long,
-    override val amountOfCoins: Long
+    override val availableStepsCount: Int,
+    override val amountOfCoins: Int
 ) : UserData {
     @Serializable
     @JvmInline
@@ -20,6 +20,6 @@ interface UserData {
     val phoneNumber: PhoneNumber
     val avatar: Avatar
     val occupation: Occupation
-    val availableStepsCount: Long
-    val amountOfCoins: Long
+    val availableStepsCount: Int
+    val amountOfCoins: Int
 }
